Refactor : Navbar 네비게이션 설정
